added the Pipeline class

+++ src/com/mojang/lib3d/lib3d.cpp
@@ -1,6 +1,6 @@
     #include <src/com/mojang/lib3d/lib3d.h>
     #include <memory.h>
-    #include <glfw3.h>
+    #include <GLFW/glfw3.h>
 
     //void lib3d::process(RenderPipeline renderPipeline, float f) {
     //    ConcurrentLinkedQueue<RenderCall> concurrentLinkedQueue = renderPipeline.getRecordingQueue();
@@ -11,7 +11,7 @@
 
     void lib3D::IntentionalCrash()
     {
-        memset((long)0L, (int)0, (long)1L);
+        memset(0, 0, 1L);
     }
 
     double lib3D::getTime()
